新增表

+++ src/main/java/com/cdkj/ylq/ao/IApplyAO.java
@@ -1,5 +1,7 @@
 package com.cdkj.ylq.ao;
 
+import java.math.BigDecimal;
+
 import com.cdkj.ylq.bo.base.Paginable;
 import com.cdkj.ylq.domain.Apply;
 import com.cdkj.ylq.dto.res.XN623020Res;
@@ -14,8 +16,8 @@ public interface IApplyAO {
     public void cancalApply(String applyUser);
 
     // 审核
-    public void doApprove(String code, String approveResult, Long sxAmount,
-            String approver, String approveNote);
+    public void doApprove(String code, String approveResult,
+            BigDecimal sxAmount, String approver, String approveNote);
 
     // 获取用户当前正在进行中的申请记录
     public Apply getCurrentApply(String userId);
